+++ app/lib/modules/hatim/hatim.dart
@@ -1,6 +1,3 @@
-export 'presentation/cubit/hatim_juz/hatim_juz_cubit.dart';
-export 'presentation/cubit/hatim_pages/hatim_pages_cubit.dart';
-export 'presentation/cubit/hatim_read/hatim_read_cubit.dart';
 export 'presentation/presentation.dart';
 export 'domain/domain.dart';
 export 'data/data.dart';
