--- app/lib/modules/hatim/presentation/cubit/hatim_juzs/hatim_juzs_state.dart
@@ -1,12 +1,0 @@
-// part of 'hatim_juzs_cubit.dart';
-
-// class HatimJuzsState extends Equatable {
-//   const HatimJuzsState({this.hatimJuzs, this.erorText, this.status = FetchStatus.loading});
-
-//   final List<HatimJus>? hatimJuzs;
-//   final String? erorText;
-//   final FetchStatus status;
-
-//   @override
-//   List<Object?> get props => [hatimJuzs, erorText, status, hatimJuzs?.map((e) => e)];
-// }
