--- app/lib/modules/hatim/presentation/cubit/hatim_pages/hatim_pages_state.dart
@@ -1,20 +1,0 @@
-// part of 'hatim_pages_cubit.dart';
-
-// class HatimPagesState extends Equatable {
-//   const HatimPagesState({this.pages, this.lastPage, this.exception});
-
-//   final List<HatimPages?>? pages;
-//   final int? lastPage;
-//   final Exception? exception;
-
-//   @override
-//   List<Object?> get props => [pages, lastPage, exception, pages?.length];
-
-//   HatimPagesState copyWith({List<HatimPages?>? pages, int? lastPage, Exception? exception}) {
-//     return HatimPagesState(
-//       pages: pages ?? this.pages,
-//       lastPage: lastPage ?? this.lastPage,
-//       exception: exception ?? this.exception,
-//     );
-//   }
-// }
