--- app/lib/modules/hatim/presentation/cubit/hatim_read/hatim_read_state.dart
@@ -1,20 +1,0 @@
-// part of 'hatim_read_cubit.dart';
-
-// class HatimReadState extends Equatable {
-//   const HatimReadState(this.status, {this.hatim, this.exception});
-
-//   final HatimReadEntity? hatim;
-//   final FetchStatus status;
-//   final Exception? exception;
-
-//   @override
-//   List<Object?> get props => [hatim, status, exception];
-
-//   HatimReadState copyWith({FetchStatus? status, HatimReadEntity? hatim, Exception? exception}) {
-//     return HatimReadState(
-//       status ?? this.status,
-//       hatim: hatim ?? this.hatim,
-//       exception: exception ?? this.exception,
-//     );
-//   }
-// }
