+++ app/lib/modules/hatim/presentation/presentation.dart
@@ -1,7 +1,6 @@
 export 'widgets/hatim_page_status_card.dart';
 export 'widgets/juz_percent_widget.dart';
 export 'widgets/vertical_text.dart';
-export 'view1/hatim_select_page_view.dart';
-export 'view1/hatim_view.dart';
-export 'cubit/hatim_juzs/hatim_juzs_cubit.dart';
+export 'view/hatim_select_page_view.dart';
+export 'view/hatim_view.dart';
 export 'bloc/hatim_bloc.dart';
