+++ app/lib/modules/hatim/presentation/view/hatim_view.dart
@@ -1,180 +1,172 @@
-// import 'package:flutter/material.dart';
-// import 'package:flutter_bloc/flutter_bloc.dart';
-// import 'package:go_router/go_router.dart';
-// import 'package:mq_ci_keys/mq_ci_keys.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:go_router/go_router.dart';
+import 'package:mq_ci_keys/mq_ci_keys.dart';
 
-// import 'package:my_quran/app/app.dart';
-// import 'package:my_quran/config/config.dart';
-// import 'package:my_quran/constants/contants.dart';
-// import 'package:my_quran/core/core.dart';
-// import 'package:my_quran/l10n/l10.dart';
-// import 'package:my_quran/modules/modules.dart';
-// import 'package:my_quran/utils/urils.dart';
+import 'package:my_quran/app/app.dart';
+import 'package:my_quran/config/config.dart';
+import 'package:my_quran/constants/contants.dart';
+import 'package:my_quran/core/core.dart';
+import 'package:my_quran/l10n/l10.dart';
+import 'package:my_quran/modules/hatim/hatim.dart';
+import 'package:my_quran/modules/modules.dart';
+import 'package:my_quran/utils/urils.dart';
 
-// class HatimView extends StatelessWidget {
-//   const HatimView({super.key});
+class HatimView extends StatelessWidget {
+  const HatimView({super.key});
 
-//   @override
-//   Widget build(BuildContext context) {
-//     return MultiBlocProvider(
-//       providers: [
-//         BlocProvider(create: (context) => HatimJuzsCubit()),
-//         BlocProvider(
-//           create: (context) => HatimReadCubit(
-//             GetHatimUseCase(
-//               HatimReadRepositoryImpl(
-//                 remoteDataSource: HatimRemoteDataSource(
-//                   token: '',
-//                   remoteClient: context.read<RemoteClient>(),
-//                 ),
-//               ),
-//             ),
-//           ),
-//         ),
-//         BlocProvider(create: (context) => HatimPagesCubit()),
-//       ],
-//       child: const HatimUI(),
-//     );
-//   }
-// }
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider(
+      create: (context) => HatimBloc(
+        remoteClient: context.read<RemoteClient>(),
+        token: context.read<AuthCubit>().state.user!.accessToken,
+      )..add(const GetHatimDashBoardEvent()),
+      child: const HatimUI(),
+    );
+  }
+}
 
-// class HatimUI extends StatefulWidget {
-//   const HatimUI({super.key});
+class HatimUI extends StatefulWidget {
+  const HatimUI({super.key});
 
-//   @override
-//   State<HatimUI> createState() => _HatimUIState();
-// }
+  @override
+  State<HatimUI> createState() => _HatimUIState();
+}
 
-// class _HatimUIState extends State<HatimUI> {
-//   @override
-//   void initState() {
-//     super.initState();
-//     getData();
-//   }
+class _HatimUIState extends State<HatimUI> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      key: const Key(MqKeys.hatimPage),
+      appBar: AppBar(
+        title: Text(context.l10n.hatim),
+        actions: [
+          BlocBuilder<HatimBloc, HatimState>(
+            buildWhen: (p, c) => p.userPagesState != c.userPagesState,
+            builder: (context, state) {
+              final subState = state.userPagesState;
+              return switch (subState) {
+                HatimUserPagesFetched() => Text('${subState.data.length}'),
+                _ => const Text('...'),
+              };
+            },
+          ),
+          const SizedBox(width: 30),
+        ],
+      ),
+      body: RefreshIndicator(
+        onRefresh: () async {
+          context.read<HatimBloc>().add(const GetHatimDashBoardEvent());
+        },
+        child: BlocConsumer<HatimBloc, HatimState>(
+          listener: (BuildContext context, HatimState state) {
+            final dashBoadrState = state.dashBoardState;
+            final eventState = state.eventState;
+            final juzsState = state.juzsState;
+            final bloc = context.read<HatimBloc>();
+            if (dashBoadrState is HatimDashBoardFetched && eventState is HatimStateInitial) {
+              bloc.add(ConnectToHatimdEvent(dashBoadrState.data.id));
+            }
+            if (eventState is HatimStateSuccess && juzsState is HatimJuzsInitial) {
+              bloc
+                ..add(GetHatimJuzsEvent(eventState.hatimId))
+                ..add(const GetHatimUserPagesEvent());
+            }
+          },
+          builder: (context, state) {
+            final juzsState = state.juzsState;
+            return switch (juzsState) {
+              HatimJuzsInitial() => const Center(child: CircularProgressIndicator()),
+              HatimJuzsLoading() => const Center(child: CircularProgressIndicator()),
+              HatimJuzsFetched() => HatimJuzListBuilder(juzsState.data),
+              HatimJuzsFailed() => Center(child: Text('${juzsState.exception}')),
+            };
+          },
+        ),
+      ),
+      floatingActionButton: BlocBuilder<HatimBloc, HatimState>(
+        builder: (context, state) {
+          final userPagesState = state.userPagesState;
+          if (userPagesState is HatimUserPagesFetched) {
+            final pages = userPagesState.data;
+            if (pages.isNotEmpty) {
+              return FloatingActionButton.extended(
+                onPressed: () async {
+                  final pageIds = pages.map((e) => e.id).toList();
+                  final pageNumbers = pages.map((e) => e.number).toList();
+                  context.read<HatimBloc>().add(SetInProgressPagesEvent(pageIds));
+                  final value = await context.pushNamed<bool>(
+                    AppRouter.hatimRead,
+                    pathParameters: {
+                      'pages': pageNumbers.toString(),
+                      'isHatim': true.toString(),
+                    },
+                  );
+                  if (value != null && value && context.mounted) {
+                    context.read<HatimBloc>().add(SetDonePagesEvent(pageIds));
+                  }
+                },
+                label: Text(context.l10n.read),
+                icon: Assets.icons.openBook.svg(),
+              );
+            }
+          }
+          return const SizedBox.shrink();
+        },
+      ),
+    );
+  }
+}
 
-//   Future<void> getData() async {
-//     final token = context.read<AuthCubit>().state.user!.accessToken;
-//     await context.read<HatimReadCubit>().getHatim(token).then(
-//       (value) {
-//         if (value != null) {
-//           context.read<HatimJuzsCubit>().connect(token, value);
-//         }
-//       },
-//     );
-//   }
+class HatimJuzListBuilder extends StatelessWidget {
+  const HatimJuzListBuilder(this.items, {super.key});
 
-//   @override
-//   Widget build(BuildContext context) {
-//     return Scaffold(
-//       key: const Key(MqKeys.hatimPage),
-//       appBar: AppBar(
-//         title: Text(context.l10n.hatim),
-//         actions: [
-//           Text(
-//             '${context.watch<HatimPagesCubit>().state.pages != null ? context.watch<HatimPagesCubit>().state.pages!.length : '...'}',
-//             style: const TextStyle(fontSize: 20),
-//           ),
-//           const SizedBox(width: 30),
-//         ],
-//       ),
-//       body: RefreshIndicator(
-//         onRefresh: () async => getData(),
-//         child: BlocBuilder<HatimJuzsCubit, HatimJuzsState>(
-//           builder: (context, state) {
-//             if (state.erorText == null && state.hatimJuzs == null) {
-//               return const Center(child: CircularProgressIndicator());
-//             } else if (state.hatimJuzs != null) {
-//               return HatimJuzListBuilder(state.hatimJuzs!);
-//             } else if (state.erorText != null) {
-//               return Center(child: Text(state.erorText!));
-//             } else {
-//               return const Center(child: Text('bilbeim'));
-//             }
-//           },
-//         ),
-//       ),
-//       floatingActionButton: context.select((HatimPagesCubit cubit) {
-//         if (cubit.state.pages != null && cubit.state.pages!.isNotEmpty) {
-//           return FloatingActionButton.extended(
-//             onPressed: () async {
-//               final token = context.read<AuthCubit>().state.user!.accessToken;
-//               final username = context.read<AuthCubit>().state.user!.username;
-//               context.read<HatimPagesCubit>().sendPage(username, token);
-//               final pages = List<int>.generate(cubit.state.pages!.length, (index) => cubit.state.pages![index]!.number)
-//                 ..sort();
-//               final value = await context.pushNamed<bool>(
-//                 AppRouter.hatimRead,
-//                 pathParameters: {
-//                   'pages': pages.toString(),
-//                   'isHatim': true.toString(),
-//                 },
-//               );
-//               if (value != null && value && context.mounted) {
-//                 context.read<HatimPagesCubit>().donePage(username, token);
-//               }
-//             },
-//             label: Text(context.l10n.read),
-//             icon: Assets.icons.openBook.svg(),
-//           );
-//         } else {
-//           return null;
-//         }
-//       }),
-//     );
-//   }
-// }
+  final List<HatimJus> items;
 
-// class HatimJuzListBuilder extends StatelessWidget {
-//   const HatimJuzListBuilder(this.items, {super.key});
+  @override
+  Widget build(BuildContext context) {
+    return ListView.builder(
+      key: const Key(MqKeys.hatimJuzsList),
+      padding: const EdgeInsets.fromLTRB(14, 20, 14, 70),
+      itemCount: items.length,
+      itemBuilder: (BuildContext context, int index) {
+        final item = items[index];
+        return Card(
+          child: Column(
+            children: [
+              ListTile(
+                key: Key(MqKeys.hatimJuzIndex(index)),
+                contentPadding: const EdgeInsets.all(8),
+                title: Text('${item.number}-${context.l10n.juz}'),
+                subtitle: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    VerticalText(context.l10n.hatimDoneRead, '${item.done}'),
+                    VerticalText(context.l10n.hatimProccessRead, '${item.inProgress}'),
+                    VerticalText(context.l10n.hatimEmptyRead, '${item.toDo}'),
+                  ],
+                ),
+                onTap: () async {
+                  final bloc = context.read<HatimBloc>();
+                  await AppBottomSheet.showBottomSheet<void>(
+                    context,
+                    scrollKey: const Key(MqKeys.hatimSelectPageScroll),
+                    initialChildSize: 0.85,
+                    BlocProvider.value(
+                      value: bloc..add(GetHatimJuzPagesEvent(item.id)),
+                      child: const HatimJusBottomSheet(),
+                    ),
+                  );
 
-//   final List<HatimJus> items;
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return ListView.builder(
-//       key: const Key(MqKeys.hatimJuzsList),
-//       padding: const EdgeInsets.fromLTRB(14, 20, 14, 70),
-//       itemCount: items.length,
-//       itemBuilder: (BuildContext context, int index) {
-//         final item = items[index];
-//         return Card(
-//           child: Column(
-//             children: [
-//               ListTile(
-//                 key: Key(MqKeys.hatimJuzIndex(index)),
-//                 contentPadding: const EdgeInsets.all(8),
-//                 title: Text('${item.number}-${context.l10n.juz}'),
-//                 subtitle: Row(
-//                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
-//                   children: [
-//                     VerticalText(context.l10n.hatimDoneRead, '${item.done}'),
-//                     VerticalText(context.l10n.hatimProccessRead, '${item.inProgress}'),
-//                     VerticalText(context.l10n.hatimEmptyRead, '${item.toDo}'),
-//                   ],
-//                 ),
-//                 onTap: () {
-//                   final token = context.read<AuthCubit>().state.user!.accessToken;
-//                   AppBottomSheet.showBottomSheet<void>(
-//                     context,
-//                     scrollKey: const Key(MqKeys.hatimSelectPageScroll),
-//                     initialChildSize: 0.85,
-//                     MultiBlocProvider(
-//                       providers: [
-//                         BlocProvider.value(value: context.read<HatimPagesCubit>()),
-//                         BlocProvider(
-//                           create: (context) => HatimJuzCubit(item.number)..connect(item.id, token),
-//                         ),
-//                       ],
-//                       child: const HatimJusBottomSheet(),
-//                     ),
-//                   );
-//                 },
-//               ),
-//               JuzPersentWidget(item),
-//             ],
-//           ),
-//         );
-//       },
-//     );
-//   }
-// }
+                  bloc.add(const ResetJuzPagesEvent());
+                },
+              ),
+              JuzPersentWidget(item),
+            ],
+          ),
+        );
+      },
+    );
+  }
+}
